requirements.txt
aiohttp==3.9.3
aiohttp-retry==2.8.3
aiosignal==1.3.1
anthropic==0.2.10
anyio==4.3.0
asttokens==2.4.1
async-timeout==4.0.3
attrs==23.2.0
azure-cognitiveservices-speech==1.36.0
blinker==1.7.0
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
clicksend-client==5.0.78
dataclasses-json==0.5.14
decorator==5.1.1
Deprecated==1.2.14
exceptiongroup==1.2.0
executing==2.0.1
fastapi==0.95.2
filelock==3.13.3
Flask==3.0.3
frozenlist==1.4.1
fsspec==2024.3.1
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
huggingface-hub==0.22.2
idna==3.6
importlib-metadata==7.0.0
ipython==8.23.0
itsdangerous==2.1.2
janus==1.0.0
jedi==0.19.1
Jinja2==3.1.3
joblib==1.3.2
langchain==0.0.149
MarkupSafe==2.1.5
marshmallow==3.21.1
matplotlib-inline==0.1.6
multidict==6.0.5
mypy-extensions==1.0.0
nltk==3.8.1
numexpr==2.10.0
numpy==1.26.4
openai==0.27.10
openapi-schema-pydantic==1.2.4
opentelemetry-api==1.24.0
opentelemetry-sdk==1.24.0
opentelemetry-semantic-conventions==0.45b0
packaging==24.0
parso==0.8.4
pexpect==4.9.0
prompt-toolkit==3.0.43
ptyprocess==0.7.0
pure-eval==0.2.2
pycparser==2.22
pydantic==1.10.15
pydub==0.25.1
Pygments==2.17.2
PyJWT==2.8.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.6
pytz==2024.1
PyYAML==6.0.1
redis==4.5.5
regex==2023.12.25
requests==2.31.0
scipy==1.13.0
six==1.16.0
sniffio==1.3.1
sounddevice==0.4.6
SQLAlchemy==2.0.29
stack-data==0.6.3
starlette==0.27.0
tenacity==8.2.3
tokenizers==0.15.2
tqdm==4.66.2
traitlets==5.14.2
twilio==8.2.0
typing-inspect==0.9.0
typing_extensions==4.11.0
urllib3==2.2.1
uvicorn==0.21.1
vocode==0.1.110
wcwidth==0.2.13
websockets==11.0.3
Werkzeug==3.0.2
wrapt==1.16.0
yarl==1.9.4
zipp==3.18.1
